requirements.txt
fastapi
uvicorn[standard]
reportlab
python-docx
PyPDF2
